+++ src/Meals/Meal.tsx
@@ -1,10 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, FlatList, Pressable } from 'react-native';
+import { View, Text, FlatList, Pressable, Alert } from 'react-native';
 import { MealItem } from '../Items'; // Adjust path if needed
 import { deleteMealItem, getDBConnection, getMealItems } from '../db-functions';
-import { styles } from '../App';
 import { useTheme } from '../Themes';
-import { useNavigation } from '@react-navigation/native';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { faX } from '@fortawesome/free-solid-svg-icons';
 
@@ -43,22 +41,50 @@ export const Meals = () => {
         renderItem={({ item }) => (
           <View style={{ padding: 10, flexDirection: 'row' }}>
             <View style={{ flex: 1 }}>
-              <Text>Id: {item.id}</Text>
-              <Text>day: {item.day}</Text>
-              <Text>type: {typeOptions[item.type]}</Text>
-              <Text>foods: {item.foods}</Text>
-              <Text>servings: {item.servings}</Text>
-              <Text>calories: {item.calories}</Text>
+              <Text
+                style={{ color: theme.h1Color, fontSize: 20, fontWeight: 500 }}
+              >
+                day: {item.day}
+              </Text>
+              <Text style={{ color: theme.h2Color, fontWeight: 500 }}>
+                type: {typeOptions[item.type]}
+              </Text>
+              <Text style={{ color: theme.h2Color, fontWeight: 500 }}>
+                foods: {item.foods}
+              </Text>
+              <Text style={{ color: theme.h2Color, fontWeight: 500 }}>
+                servings: {item.servings}
+              </Text>
+              <Text style={{ color: theme.h2Color, fontWeight: 500 }}>
+                calories: {item.calories}
+              </Text>
             </View>
             <Pressable
-              onPress={async () => {
-                const db = await getDBConnection();
-                await deleteMealItem(db, item.id);
-                const Meals = await getMealItems(db);
-                setMeals(Meals);
+              onPress={() => {
+                Alert.alert(
+                  'Are you sure?',
+                  'This cannot be undone.\n You will never see this again.',
+                  [
+                    {
+                      text: 'Cancel',
+                      onPress: () => console.log('Cancel Pressed'),
+                      style: 'cancel',
+                    },
+                    {
+                      text: 'OK',
+                      onPress: async () => {
+                        const db = await getDBConnection();
+                        await deleteMealItem(db, item.id);
+                        const Meals = await getMealItems(db);
+                        setMeals(Meals);
+                      },
+                    },
+                  ],
+                  { cancelable: true },
+                );
               }}
             >
-              <FontAwesomeIcon icon={faX} />
+              <FontAwesomeIcon icon={faX} color={theme.h1Color} />
             </Pressable>
           </View>
         )}
